Remove TAEF_ARCH which is useless after remove select_filter. Move strip_dxil_validator_path to TestingConfig.
