+++ cmake/modules/FindTAEF.cmake
@@ -42,21 +42,14 @@ set(TAEF_NUGET_BIN ${TAEF_INCLUDE_DIR}/../Binaries/Release)
 set(TAEF_SDK_BIN ${TAEF_INCLUDE_DIR}/../../Runtimes/TAEF)
 
 if ((CMAKE_GENERATOR_PLATFORM STREQUAL "x64") OR ("${CMAKE_C_COMPILER_ARCHITECTURE_ID}" STREQUAL "x64"))
-  set(TAEF_ARCH "x64")
+  set(TAEF_BIN_ARCH "amd64")
 elseif ((CMAKE_GENERATOR_PLATFORM STREQUAL "x86") OR ("${CMAKE_C_COMPILER_ARCHITECTURE_ID}" STREQUAL "x86"))
-  set(TAEF_ARCH "x86")
+  set(TAEF_BIN_ARCH "x86")
 elseif ((CMAKE_GENERATOR_PLATFORM MATCHES "ARM64.*") OR ("${CMAKE_C_COMPILER_ARCHITECTURE_ID}" MATCHES "ARM64.*"))
-  set(TAEF_ARCH "arm64")
+  set(TAEF_BIN_ARCH "arm64")
 elseif ((CMAKE_GENERATOR_PLATFORM MATCHES "ARM.*") OR ("${CMAKE_C_COMPILER_ARCHITECTURE_ID}" MATCHES "ARM.*"))
-  set(TAEF_ARCH "arm")
+  set(TAEF_BIN_ARCH "arm")
 endif((CMAKE_GENERATOR_PLATFORM STREQUAL "x64") OR ("${CMAKE_C_COMPILER_ARCHITECTURE_ID}" STREQUAL "x64"))
-
-set (TAEF_ARCH ${TAEF_ARCH} CACHE INTERNAL "arch for taef test")
-
-set (TAEF_BIN_ARCH ${TAEF_ARCH})
-if (TAEF_BIN_ARCH STREQUAL "x64")
-  set (TAEF_BIN_ARCH "amd64")
-endif()
 
 find_program(TAEF_EXECUTABLE te.exe PATHS
   ${CMAKE_SOURCE_DIR}/external/taef/build/Binaries/${TAEF_BIN_ARCH}
