+++ projects/dxilconv/test/taef/lit.cfg
@@ -37,11 +37,5 @@ verbose = lit_config.params.get('verbose', None)
 if verbose == None:
     extra_params.append('/logOutput:LowWithConsoleBuffering')
 
-arch = getattr(config, 'taef_arch', None)
-arch_filter = str.format("@Architecture='{}'", arch)
-select_filter = str.format("/select:\"{}\"", arch_filter)
-extra_params.append(select_filter)
-
-need_dxil_validator = False
-config.test_format = lit.formats.TaefTest(config.te, test_dll, test_dir, extra_params, need_dxil_validator)
+config.test_format = lit.formats.TaefTest(config.te, test_dll, test_dir, extra_params)
