+++ projects/dxilconv/test/taef/lit.site.cfg.in
@@ -6,7 +6,6 @@ config.llvm_src_root = "@LLVM_SOURCE_DIR@"
 config.llvm_obj_root = "@LLVM_BINARY_DIR@"
 config.llvm_build_mode = "@LLVM_BUILD_MODE@"
 config.te = "@TAEF_EXECUTABLE@"
-config.taef_arch = "@TAEF_ARCH@"
 
 # Support substitution of the tools_dir, libs_dirs, and build_mode with user
 # parameters. This is used when we can't determine the tool dir at
