+++ tools/clang/test/taef/lit.cfg
@@ -41,5 +41,5 @@ else:
   verbose = lit_config.params.get('verbose', None)
   if verbose == None:
       extra_params.append('/logOutput:LowWithConsoleBuffering')
-  need_dxil_validator = False
-  config.test_format = lit.formats.TaefTest(config.te, test_dll, test_dir, extra_params, need_dxil_validator)
+
+  config.test_format = lit.formats.TaefTest(config.te, test_dll, test_dir, extra_params)
