+++ tools/clang/test/taef_exec/lit.cfg
@@ -58,5 +58,4 @@ else:
   if taef_extra_params:
     extra_params.extend(taef_extra_params.split(';'))
 
-  need_dxil_validator = True
-  config.test_format = lit.formats.TaefTest(config.te, test_dll, test_dir, extra_params, need_dxil_validator)
+  config.test_format = lit.formats.TaefTest(config.te, test_dll, test_dir, extra_params)
