+++ utils/lit/lit/formats/taef.py
@@ -8,12 +8,11 @@ import lit.util
 from .base import TestFormat
 
 class TaefTest(TestFormat):
-    def __init__(self, te_path, test_dll, test_path, extra_params, need_dxil_validator):
+    def __init__(self, te_path, test_dll, test_path, extra_params):
         self.te = te_path
         self.test_dll = test_dll
         self.test_path = test_path
         self.extra_params = extra_params
-        self.need_dxil_validator = need_dxil_validator
         # NOTE: when search test, always running on test_dll,
         #       use test_searched to make sure only add test once.
         #       If TaeftTest is created in directory with sub directory,
